left - allow environment variable access in policy interpolation
